+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### New Features
 
+- Add document history navigation to the toolbar for folder documents.
 - Add the Customize Syntax Menu command to the Available syntaxes action menu in Format settings to hide specific syntaxes from syntax menus.